+++ app/main.js
@@ -1,17 +1,33 @@
-// process.on('uncaughtException', (...args) => console.error(...args))
-const { app, nativeTheme, BrowserWindow, Tray, Menu, ipcMain, shell, dialog, globalShortcut } = require('electron')
+const {
+  app, nativeTheme, BrowserWindow, Tray, Menu, ipcMain,
+  shell, dialog, globalShortcut
+} = require('electron')
 const path = require('path')
 const i18next = require('i18next')
 const Backend = require('i18next-node-fs-backend')
 const log = require('electron-log')
+const Store = require('electron-store')
 
-startI18next()
+process.on('uncaughtException', (err, _) => {
+  log.error(err)
+  const dialogOpts = {
+    type: 'error',
+    title: 'Stretchly',
+    message: 'An error occured while running Stretchly and it will now quit. To report the issue, click Report.',
+    buttons: ['Report', 'OK']
+  }
+  dialog.showMessageBox(dialogOpts).then((returnValue) => {
+    if (returnValue.response === 0) {
+      shell.openExternal('https://github.com/hovancik/stretchly/issues')
+    }
+    app.quit()
+  })
+})
 
 nativeTheme.on('updated', function theThemeHasChanged () {
   appIcon.setImage(trayIconPath())
 })
 
-const AppSettings = require('./utils/settings')
 const Utils = require('./utils/utils')
 const IdeasLoader = require('./utils/ideasLoader')
 const BreaksPlanner = require('./breaksPlanner')
@@ -34,6 +50,10 @@ let myStretchlyWindow = null
 let settings
 let pausedForSuspendOrLock = false
 let nextIdea = null
+let appIsQuitting = false
+let updateChecker
+
+require('@electron/remote/main').initialize()
 
 app.setAppUserModelId('net.hovancik.stretchly')
 
@@ -50,23 +70,105 @@ if (!gotTheLock) {
   const cmd = new Command(args, app.getVersion())
   cmd.runOrForward()
   app.quit()
-  return
 }
 
-app.on('ready', startProcessWin)
-app.on('ready', loadSettings)
-app.on('ready', createTrayIcon)
-app.on('ready', startPowerMonitoring)
+app.on('ready', initialize)
 app.on('second-instance', runCommand)
 app.on('window-all-closed', () => {
   // do nothing, so app wont get closed
 })
+app.on('before-quit', () => {
+  appIsQuitting = true
+})
+
+function initialize (isAppStart = true) {
+  // TODO maybe we should not reinitialize but handle everything when we save new values for preferences
+  log.info(`Stretchly: ${isAppStart ? '' : 're'}initializing...`)
+  require('events').defaultMaxListeners = 200 // for watching Store changes
+  if (!settings) {
+    settings = new Store({ defaults: require('./utils/defaultSettings'), watch: true })
+    log.info('Stretchly: loading preferences')
+    Store.initRenderer()
+    Object.entries(settings.store).forEach(([key, _]) => {
+      settings.onDidChange(key, (newValue, oldValue) => {
+        log.info(`Stretchly: setting '${key}' to '${newValue}' (was '${oldValue}')`)
+      })
+    })
+  }
+  if (!breakPlanner) {
+    breakPlanner = new BreaksPlanner(settings)
+    breakPlanner.nextBreak()
+    breakPlanner.on('startMicrobreakNotification', () => { startMicrobreakNotification() })
+    breakPlanner.on('startBreakNotification', () => { startBreakNotification() })
+    breakPlanner.on('startMicrobreak', () => { startMicrobreak() })
+    breakPlanner.on('finishMicrobreak', (shouldPlaySound) => { finishMicrobreak(shouldPlaySound) })
+    breakPlanner.on('startBreak', () => { startBreak() })
+    breakPlanner.on('finishBreak', (shouldPlaySound) => { finishBreak(shouldPlaySound) })
+    breakPlanner.on('resumeBreaks', () => { resumeBreaks() })
+    breakPlanner.on('updateToolTip', function () {
+      updateTray()
+    })
+  } else {
+    breakPlanner.clear()
+    breakPlanner.appExclusionsManager.reinitialize(settings)
+    breakPlanner.doNotDisturb(settings.get('monitorDnd'))
+    breakPlanner.naturalBreaks(settings.get('naturalBreaks'))
+    breakPlanner.nextBreak()
+  }
+  if (!appIcon) {
+    if (process.platform === 'darwin') {
+      app.dock.hide()
+    }
+    appIcon = new Tray(trayIconPath())
+  }
+  startI18next()
+  setInterval(updateTray, 10000)
+  startProcessWin()
+  createWelcomeWindow()
+  nativeTheme.themeSource = settings.get('themeSource')
+
+  require('fs').readFile(path.join(app.getPath('userData'), 'stamp'), 'utf8', (err, data) => {
+    if (err) {
+      return
+    }
+    const { DateTime } = require('luxon')
+    if (DateTime.fromISO(data).month === DateTime.now().month) {
+      global.shared.isContributor = true
+      log.info('Stretchly: Thanks for your contributions!')
+      if (preferencesWin) {
+        preferencesWin.send('enableContributorPreferences')
+      }
+      updateTray()
+    }
+  })
+  startPowerMonitoring()
+  if (preferencesWin) {
+    preferencesWin.send('renderSettings', settingsToSend())
+  }
+  if (welcomeWin) {
+    welcomeWin.send('renderSettings', settingsToSend())
+  }
+  if (contributorPreferencesWindow) {
+    contributorPreferencesWindow.send('renderSettings', settingsToSend())
+  }
+  globalShortcut.unregisterAll()
+  if (settings.get('resumeBreaksShortcut') !== '') {
+    globalShortcut.register(settings.get('resumeBreaksShortcut'), () => {
+      resumeBreaks(false)
+    })
+  }
+  if (settings.get('pauseBreaksShortcut') !== '') {
+    globalShortcut.register(settings.get('pauseBreaksShortcut'), () => {
+      pauseBreaks(1)
+    })
+  }
+}
 
 function startI18next () {
   i18next
     .use(Backend)
     .init({
-      lng: 'en',
+      lng: settings.get('language'),
       fallbackLng: 'en',
       debug: false,
       backend: {
@@ -92,9 +194,11 @@ i18next.on('languageChanged', function (lng) {
 function onSuspendOrLock () {
   log.info('System: suspend or lock')
   if (!breakPlanner.isPaused) {
-    pausedForSuspendOrLock = true
-    pauseBreaks(1)
-    updateTray()
+    if (settings.get('pauseForSuspendOrLock')) {
+      pausedForSuspendOrLock = true
+      pauseBreaks(1)
+      updateTray()
+    }
   }
 }
 
@@ -103,7 +207,7 @@ function onResumeOrUnlock () {
   if (pausedForSuspendOrLock) {
     pausedForSuspendOrLock = false
     resumeBreaks(false)
-  } else if (breakPlanner.isPaused) {
+  } else {
     // corrrect the planner for the time spent in suspend
     breakPlanner.correctScheduler()
   }
@@ -124,9 +228,9 @@ function numberOfDisplays () {
 }
 
 function closeWindows (windowArray) {
-  for (let i = windowArray.length - 1; i >= 0; i--) {
-    windowArray[i].hide()
-    windowArray[i].close()
+  for (const window of windowArray) {
+    window.hide()
+    window.close()
   }
   return null
 }
@@ -134,11 +238,21 @@ function closeWindows (windowArray) {
 function displaysX (displayID = -1, width = 800, fullscreen = false) {
   const electron = require('electron')
   let theScreen
+
+  if (!settings.get('allScreens')) {
+    if (settings.get('screen') === 'primary') {
+      theScreen = electron.screen.getPrimaryDisplay()
+    } else if (settings.get('screen') === 'cursor') {
+      theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
+    } else {
+      displayID = parseInt(settings.get('screen'))
+    }
+  }
+
   if (displayID === -1) {
     theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
   } else if (displayID >= numberOfDisplays() || displayID < 0) {
-    // Graceful handling of invalid displayID
-    log.warn('Stretchly: invalid displayID to displaysX')
+    log.warn(`Stretchly: invalid displayID ${displayID} to displaysX`)
     theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
   } else {
     const screens = electron.screen.getAllDisplays()
@@ -155,11 +269,21 @@ function displaysX (displayID = -1, width = 800, fullscreen = false) {
 function displaysY (displayID = -1, height = 600, fullscreen = false) {
   const electron = require('electron')
   let theScreen
+
+  if (!settings.get('allScreens')) {
+    if (settings.get('screen') === 'primary') {
+      theScreen = electron.screen.getPrimaryDisplay()
+    } else if (settings.get('screen') === 'cursor') {
+      theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
+    } else {
+      displayID = parseInt(settings.get('screen'))
+    }
+  }
+
   if (displayID === -1) {
     theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
-  } else if (displayID >= numberOfDisplays()) {
-    // Graceful handling of invalid displayID
-    log.warn('Stretchly: invalid displayID to displaysY')
+  } else if (displayID >= numberOfDisplays() || displayID < 0) {
+    log.warn(`Stretchly: invalid displayID ${displayID} to displaysY`)
     theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
   } else {
     const screens = electron.screen.getAllDisplays()
@@ -176,11 +300,21 @@ function displaysY (displayID = -1, height = 600, fullscreen = false) {
 function displaysWidth (displayID = -1) {
   const electron = require('electron')
   let theScreen
+
+  if (!settings.get('allScreens')) {
+    if (settings.get('screen') === 'primary') {
+      theScreen = electron.screen.getPrimaryDisplay()
+    } else if (settings.get('screen') === 'cursor') {
+      theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
+    } else {
+      displayID = parseInt(settings.get('screen'))
+    }
+  }
+
   if (displayID === -1) {
     theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
-  } else if (displayID >= numberOfDisplays()) {
-    // Graceful handling of invalid displayID
-    log.warn('Stretchly: invalid displayID to displaysY')
+  } else if (displayID >= numberOfDisplays() || displayID < 0) {
+    log.warn(`Stretchly: invalid displayID ${displayID} to displaysWidth`)
     theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
   } else {
     const screens = electron.screen.getAllDisplays()
@@ -193,11 +327,21 @@ function displaysWidth (displayID = -1) {
 function displaysHeight (displayID = -1) {
   const electron = require('electron')
   let theScreen
+
+  if (!settings.get('allScreens')) {
+    if (settings.get('screen') === 'primary') {
+      theScreen = electron.screen.getPrimaryDisplay()
+    } else if (settings.get('screen') === 'cursor') {
+      theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
+    } else {
+      displayID = parseInt(settings.get('screen'))
+    }
+  }
+
   if (displayID === -1) {
     theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
-  } else if (displayID >= numberOfDisplays()) {
-    // Graceful handling of invalid displayID
-    log.warn('Stretchly: invalid displayID to displaysY')
+  } else if (displayID >= numberOfDisplays() || displayID < 0) {
+    log.warn(`Stretchly: invalid displayID ${displayID} to displaysHeight`)
     theScreen = electron.screen.getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
   } else {
     const screens = electron.screen.getAllDisplays()
@@ -207,18 +351,11 @@ function displaysHeight (displayID = -1) {
   return Math.ceil(bounds.height)
 }
 
-function createTrayIcon () {
-  if (process.platform === 'darwin') {
-    app.dock.hide()
-  }
-  appIcon = new Tray(trayIconPath())
-  updateTray()
-  setInterval(updateTray, 10000)
-}
-
 function trayIconPath () {
   const params = {
-    paused: breakPlanner.isPaused || breakPlanner.dndManager.isOnDnd || breakPlanner.naturalBreaksManager.isSchedulerCleared,
+    paused: breakPlanner.isPaused || breakPlanner.dndManager.isOnDnd ||
+      breakPlanner.naturalBreaksManager.isSchedulerCleared ||
+      breakPlanner.appExclusionsManager.isSchedulerCleared,
     monochrome: settings.get('useMonochromeTrayIcon'),
     inverted: settings.get('useMonochromeInvertedTrayIcon'),
     darkMode: nativeTheme.shouldUseDarkColors,
@@ -230,7 +367,7 @@ function trayIconPath () {
 
 function windowIconPath () {
   const params = {
-    paused: breakPlanner.isPaused,
+    paused: false,
     monochrome: settings.get('useMonochromeTrayIcon'),
     inverted: settings.get('useMonochromeInvertedTrayIcon'),
     darkMode: nativeTheme.shouldUseDarkColors,
@@ -241,11 +378,17 @@ function windowIconPath () {
 }
 
 function startProcessWin () {
+  if (processWin) {
+    planVersionCheck()
+    return
+  }
   const modalPath = path.join('file://', __dirname, '/process.html')
   processWin = new BrowserWindow({
     show: false,
+    backgroundThrottling: false,
     webPreferences: {
-      nodeIntegration: true
+      preload: path.join(__dirname, './process.js'),
+      enableRemoteModule: true
     }
   })
   processWin.loadURL(modalPath)
@@ -254,19 +397,20 @@ function startProcessWin () {
   })
 }
 
-function createWelcomeWindow () {
-  if (settings.get('isFirstRun')) {
+function createWelcomeWindow (isAppStart = true) {
+  if (settings.get('isFirstRun') && isAppStart) {
     const modalPath = path.join('file://', __dirname, '/welcome.html')
     welcomeWin = new BrowserWindow({
       x: displaysX(-1, 1000),
-      y: displaysY(),
+      y: displaysY(-1, 770),
       width: 1000,
-      height: 685,
+      height: 770,
       autoHideMenuBar: true,
       icon: windowIconPath(),
       backgroundColor: 'EDEDED',
       webPreferences: {
-        nodeIntegration: true
+        preload: path.join(__dirname, './welcome.js'),
+        enableRemoteModule: true
       }
     })
     welcomeWin.loadURL(modalPath)
@@ -292,7 +436,8 @@ function createContributorSettingsWindow () {
     icon: windowIconPath(),
     backgroundColor: 'EDEDED',
     webPreferences: {
-      nodeIntegration: true
+      preload: path.join(__dirname, './contributor-preferences.js'),
+      enableRemoteModule: true
     }
   })
   contributorPreferencesWindow.loadURL(modalPath)
@@ -320,7 +465,7 @@ function createSyncPreferencesWindow () {
     backgroundColor: 'whitesmoke',
     webPreferences: {
       preload: path.resolve(__dirname, './electron-bridge.js'),
-      nodeIntegration: false
+      enableRemoteModule: true
     }
   })
   syncPreferencesWindow.loadURL(syncPreferencesUrl)
@@ -332,7 +477,11 @@ function createSyncPreferencesWindow () {
 }
 
 function planVersionCheck (seconds = 1) {
-  setTimeout(checkVersion, seconds * 1000)
+  if (updateChecker) {
+    clearInterval(updateChecker)
+    updateChecker = null
+  }
+  updateChecker = setTimeout(checkVersion, seconds * 1000)
 }
 
 function checkVersion () {
@@ -377,16 +526,19 @@ function startMicrobreak () {
   const postponableDurationPercent = settings.get('microbreakPostponableDurationPercent')
   const postponable = settings.get('microbreakPostpone') &&
     breakPlanner.postponesNumber < postponesLimit && postponesLimit > 0
+  const showBreaksAsRegularWindows = settings.get('showBreaksAsRegularWindows')
 
   if (!strictMode || postponable) {
-    globalShortcut.register(settings.get('endBreakShortcut'), () => {
-      const passedPercent = (Date.now() - startTime) / breakDuration * 100
-      if (Utils.canPostpone(postponable, passedPercent, postponableDurationPercent)) {
-        postponeMicrobreak()
-      } else if (Utils.canSkip(strictMode, postponable, passedPercent, postponableDurationPercent)) {
-        finishMicrobreak(false)
-      }
-    })
+    if (settings.get('endBreakShortcut') !== '') {
+      globalShortcut.register(settings.get('endBreakShortcut'), () => {
+        const passedPercent = (Date.now() - startTime) / breakDuration * 100
+        if (Utils.canPostpone(postponable, passedPercent, postponableDurationPercent)) {
+          postponeMicrobreak()
+        } else if (Utils.canSkip(strictMode, postponable, passedPercent, postponableDurationPercent)) {
+          finishMicrobreak(false)
+        }
+      })
+    }
   }
 
   const modalPath = path.join('file://', __dirname, '/microbreak.html')
@@ -406,16 +558,18 @@ function startMicrobreak () {
       autoHideMenuBar: true,
       icon: windowIconPath(),
       resizable: false,
-      frame: false,
+      frame: showBreaksAsRegularWindows,
       show: false,
+      backgroundThrottling: false,
       transparent: settings.get('transparentMode'),
       backgroundColor: calculateBackgroundColor(),
-      skipTaskbar: true,
-      focusable: false,
+      skipTaskbar: !showBreaksAsRegularWindows,
+      focusable: showBreaksAsRegularWindows,
+      alwaysOnTop: !showBreaksAsRegularWindows,
       title: 'Stretchly',
-      alwaysOnTop: true,
       webPreferences: {
-        nodeIntegration: true
+        preload: path.join(__dirname, './microbreak.js'),
+        enableRemoteModule: true
       }
     }
 
@@ -434,10 +588,19 @@ function startMicrobreak () {
     microbreakWinLocal.setSize(windowOptions.width, windowOptions.height)
     // microbreakWinLocal.webContents.openDevTools()
     microbreakWinLocal.once('ready-to-show', () => {
-      microbreakWinLocal.showInactive()
+      if (showBreaksAsRegularWindows) {
+        microbreakWinLocal.show()
+      } else {
+        microbreakWinLocal.showInactive()
+      }
+
       log.info(`Stretchly: showing window ${localDisplayId + 1} of ${numberOfDisplays()}`)
       if (process.platform === 'darwin') {
-        microbreakWinLocal.setKiosk(settings.get('fullscreen'))
+        if (showBreaksAsRegularWindows) {
+          microbreakWinLocal.setFullScreen(settings.get('fullscreen'))
+        } else {
+          microbreakWinLocal.setKiosk(settings.get('fullscreen'))
+        }
       }
       if (localDisplayId === 0) {
         breakPlanner.emit('microbreakStarted', true)
@@ -445,7 +608,7 @@ function startMicrobreak () {
       }
       microbreakWinLocal.webContents.send('microbreakIdea', idea)
       microbreakWinLocal.webContents.send('progress', startTime,
-        breakDuration, strictMode, postponable, postponableDurationPercent, settings.get('endBreakShortcut'))
+        breakDuration, strictMode, postponable, postponableDurationPercent)
       if (!settings.get('fullscreen') && process.platform !== 'darwin') {
         setTimeout(() => {
           microbreakWinLocal.center()
@@ -455,8 +618,15 @@ function startMicrobreak () {
 
     microbreakWinLocal.loadURL(modalPath)
     microbreakWinLocal.setVisibleOnAllWorkspaces(true)
-    microbreakWinLocal.setAlwaysOnTop(true, 'screen-saver')
+    microbreakWinLocal.setAlwaysOnTop(!showBreaksAsRegularWindows, 'pop-up-menu')
     if (microbreakWinLocal) {
+      microbreakWinLocal.on('close', (e) => {
+        if (settings.get('showBreaksAsRegularWindows')) {
+          if (!appIsQuitting && !microbreakWinLocal.fullScreen) {
+            e.preventDefault()
+          }
+        }
+      })
       microbreakWinLocal.on('closed', () => {
         microbreakWinLocal = null
       })
@@ -464,8 +634,14 @@ function startMicrobreak () {
     microbreakWins.push(microbreakWinLocal)
 
     if (!settings.get('allScreens')) {
+      if (numberOfDisplays() > 1) {
+        log.info('Stretchly: not showing on more Monitors as it is disabled.')
+      }
       break
     }
+  }
+  if (process.platform === 'darwin') {
+    app.dock.hide()
   }
   updateTray()
 }
@@ -486,16 +662,19 @@ function startBreak () {
   const postponableDurationPercent = settings.get('breakPostponableDurationPercent')
   const postponable = settings.get('breakPostpone') &&
     breakPlanner.postponesNumber < postponesLimit && postponesLimit > 0
+  const showBreaksAsRegularWindows = settings.get('showBreaksAsRegularWindows')
 
   if (!strictMode || postponable) {
-    globalShortcut.register(settings.get('endBreakShortcut'), () => {
-      const passedPercent = (Date.now() - startTime) / breakDuration * 100
-      if (Utils.canPostpone(postponable, passedPercent, postponableDurationPercent)) {
-        postponeBreak()
-      } else if (Utils.canSkip(strictMode, postponable, passedPercent, postponableDurationPercent)) {
-        finishBreak(false)
-      }
-    })
+    if (settings.get('endBreakShortcut') !== '') {
+      globalShortcut.register(settings.get('endBreakShortcut'), () => {
+        const passedPercent = (Date.now() - startTime) / breakDuration * 100
+        if (Utils.canPostpone(postponable, passedPercent, postponableDurationPercent)) {
+          postponeBreak()
+        } else if (Utils.canSkip(strictMode, postponable, passedPercent, postponableDurationPercent)) {
+          finishBreak(false)
+        }
+      })
+    }
   }
 
   const modalPath = path.join('file://', __dirname, '/break.html')
@@ -516,16 +695,18 @@ function startBreak () {
       autoHideMenuBar: true,
       icon: windowIconPath(),
       resizable: false,
-      frame: false,
+      frame: showBreaksAsRegularWindows,
       show: false,
+      backgroundThrottling: false,
       transparent: settings.get('transparentMode'),
       backgroundColor: calculateBackgroundColor(),
-      skipTaskbar: true,
-      focusable: false,
+      skipTaskbar: !showBreaksAsRegularWindows,
+      focusable: showBreaksAsRegularWindows,
+      alwaysOnTop: !showBreaksAsRegularWindows,
       title: 'Stretchly',
-      alwaysOnTop: true,
       webPreferences: {
-        nodeIntegration: true
+        preload: path.join(__dirname, './break.js'),
+        enableRemoteModule: true
       }
     }
 
@@ -544,10 +725,19 @@ function startBreak () {
     breakWinLocal.setSize(windowOptions.width, windowOptions.height)
     // breakWinLocal.webContents.openDevTools()
     breakWinLocal.once('ready-to-show', () => {
-      breakWinLocal.showInactive()
+      if (showBreaksAsRegularWindows) {
+        breakWinLocal.show()
+      } else {
+        breakWinLocal.showInactive()
+      }
+
       log.info(`Stretchly: showing window ${localDisplayId + 1} of ${numberOfDisplays()}`)
       if (process.platform === 'darwin') {
-        breakWinLocal.setKiosk(settings.get('fullscreen'))
+        if (showBreaksAsRegularWindows) {
+          breakWinLocal.setFullScreen(settings.get('fullscreen'))
+        } else {
+          breakWinLocal.setKiosk(settings.get('fullscreen'))
+        }
       }
       if (localDisplayId === 0) {
         breakPlanner.emit('breakStarted', true)
@@ -555,7 +745,7 @@ function startBreak () {
       }
       breakWinLocal.webContents.send('breakIdea', idea)
       breakWinLocal.webContents.send('progress', startTime,
-        breakDuration, strictMode, postponable, postponableDurationPercent, settings.get('endBreakShortcut'))
+        breakDuration, strictMode, postponable, postponableDurationPercent)
       if (!settings.get('fullscreen') && process.platform !== 'darwin') {
         setTimeout(() => {
           breakWinLocal.center()
@@ -564,8 +754,15 @@ function startBreak () {
     })
     breakWinLocal.loadURL(modalPath)
     breakWinLocal.setVisibleOnAllWorkspaces(true)
-    breakWinLocal.setAlwaysOnTop(true, 'screen-saver')
+    breakWinLocal.setAlwaysOnTop(!showBreaksAsRegularWindows, 'pop-up-menu')
     if (breakWinLocal) {
+      breakWinLocal.on('close', (e) => {
+        if (settings.get('showBreaksAsRegularWindows')) {
+          if (!appIsQuitting && !breakWinLocal.fullScreen) {
+            e.preventDefault()
+          }
+        }
+      })
       breakWinLocal.on('closed', () => {
         breakWinLocal = null
       })
@@ -573,15 +770,22 @@ function startBreak () {
     breakWins.push(breakWinLocal)
 
     if (!settings.get('allScreens')) {
+      if (numberOfDisplays() > 1) {
+        log.info('Stretchly: not showing on more Monitors as it is disabled.')
+      }
       break
     }
   }
-
+  if (process.platform === 'darwin') {
+    app.dock.hide()
+  }
   updateTray()
 }
 
 function breakComplete (shouldPlaySound, windows) {
-  globalShortcut.unregister(settings.get('endBreakShortcut'))
+  if (globalShortcut.isRegistered(settings.get('endBreakShortcut'))) {
+    globalShortcut.unregister(settings.get('endBreakShortcut'))
+  }
   if (shouldPlaySound && !settings.get('silentNotifications')) {
     processWin.webContents.send('playSound', settings.get('audio'), settings.get('volume'))
   }
@@ -662,27 +866,6 @@ function calculateBackgroundColor () {
   return '#' + Math.round(opacity * 255).toString(16) + themeColor.substr(1)
 }
 
-function loadSettings () {
-  const dir = app.getPath('userData')
-  const settingsFile = `${dir}/config.json`
-  settings = new AppSettings(settingsFile)
-  breakPlanner = new BreaksPlanner(settings)
-  breakPlanner.nextBreak() // plan first break
-  breakPlanner.on('startMicrobreakNotification', () => { startMicrobreakNotification() })
-  breakPlanner.on('startBreakNotification', () => { startBreakNotification() })
-  breakPlanner.on('startMicrobreak', () => { startMicrobreak() })
-  breakPlanner.on('finishMicrobreak', (shouldPlaySound) => { finishMicrobreak(shouldPlaySound) })
-  breakPlanner.on('startBreak', () => { startBreak() })
-  breakPlanner.on('finishBreak', (shouldPlaySound) => { finishBreak(shouldPlaySound) })
-  breakPlanner.on('resumeBreaks', () => { resumeBreaks() })
-  breakPlanner.on('updateToolTip', function () {
-    updateTray()
-  })
-  i18next.changeLanguage(settings.get('language'))
-  createWelcomeWindow()
-  nativeTheme.themeSource = settings.get('themeSource')
-}
-
 function loadIdeas () {
   let breakIdeasData
   let microbreakIdeasData
@@ -711,7 +894,7 @@ function pauseBreaks (milliseconds) {
 
 function resumeBreaks (notify = true) {
   if (breakPlanner.dndManager.isOnDnd) {
-    log.info('Stretchly: not resuming breaks because in DND')
+    log.info('Stretchly: not resuming breaks because in Do Not Disturb')
   } else {
     breakPlanner.resume()
     log.info('Stretchly: resuming breaks')
@@ -729,9 +912,9 @@ function createPreferencesWindow () {
     return
   }
   const modalPath = path.join('file://', __dirname, '/preferences.html')
-  const maxHeight = (electron.screen
+  const maxHeight = electron.screen
     .getDisplayNearestPoint(electron.screen.getCursorScreenPoint())
-    .workAreaSize.height - 530) / 2.0 + 490
+    .workAreaSize.height * 0.9
   preferencesWin = new BrowserWindow({
     autoHideMenuBar: true,
     icon: windowIconPath(),
@@ -742,7 +925,8 @@ function createPreferencesWindow () {
     y: displaysY(-1, 530),
     backgroundColor: '#EDEDED',
     webPreferences: {
-      nodeIntegration: true
+      preload: path.join(__dirname, './preferences.js'),
+      enableRemoteModule: true
     }
   })
   preferencesWin.loadURL(modalPath)
@@ -759,7 +943,6 @@ function updateTray () {
 
 function getTrayMenu () {
   const trayMenu = []
-  const doNotDisturb = breakPlanner.dndManager.isOnDnd
 
   if (global.shared.isNewVersion) {
     trayMenu.push({
@@ -792,7 +975,7 @@ function getTrayMenu () {
     })
   }
 
-  if (!(breakPlanner.isPaused || breakPlanner.dndManager.isOnDnd)) {
+  if (!(breakPlanner.isPaused || breakPlanner.dndManager.isOnDnd || breakPlanner.appExclusionsManager.isSchedulerCleared)) {
     let submenu = []
     if (settings.get('microbreak')) {
       submenu = submenu.concat([{
@@ -822,7 +1005,7 @@ function getTrayMenu () {
         updateTray()
       }
     })
-  } else if (!doNotDisturb) {
+  } else if (!(breakPlanner.dndManager.isOnDnd || breakPlanner.appExclusionsManager.isSchedulerCleared)) {
     trayMenu.push({
       label: i18next.t('main.pause'),
       submenu: [
@@ -856,6 +1039,11 @@ function getTrayMenu () {
     // nothing
   } else if (breakPlanner.scheduler.reference === 'finishBreak' && settings.get('breakStrictMode')) {
     // nothing
+  } else if (!(breakPlanner.dndManager.isOnDnd || breakPlanner.appExclusionsManager.isSchedulerCleared)) {
+    trayMenu.push({
+      label: i18next.t('main.resetBreaks'),
+      click: resetBreaks
+    })
   }
 
   trayMenu.push({
@@ -1020,10 +1208,10 @@ ipcMain.on('restore-defaults', (event) => {
   }
   dialog.showMessageBox(dialogOpts).then((returnValue) => {
     if (returnValue.response === 0) {
-      settings.restoreDefaults()
-      i18next.changeLanguage(settings.get('language'))
-      updateTray()
-      event.sender.webContents.send('renderSettings', settingsToSend())
+      log.info('Stretchly: restoring default settings')
+      settings.store = Object.assign(require('./utils/defaultSettings'), { isFirstRun: false })
+      initialize(false)
+      event.sender.send('renderSettings', settingsToSend())
     }
   })
 })
@@ -1035,7 +1223,7 @@ ipcMain.on('send-settings', function (event) {
 function settingsToSend () {
   const loginItemSettings = app.getLoginItemSettings()
   const openAtLogin = loginItemSettings.openAtLogin
-  return Object.assign({}, settings.data, { openAtLogin: openAtLogin })
+  return Object.assign({}, settings.store, { openAtLogin: openAtLogin })
 }
 
 ipcMain.on('play-sound', function (event, sound) {
@@ -1048,8 +1236,7 @@ ipcMain.on('show-debug', function (event) {
   const breaknumber = breakPlanner.breakNumber
   const postponesnumber = breakPlanner.postponesNumber
   const doNotDisturb = breakPlanner.dndManager.isOnDnd
-  const dir = app.getPath('userData')
-  const settingsFile = path.join(dir, 'config.json')
+  const settingsFile = settings.path
   const logsFile = log.transports.file.getFile().path
   event.sender.send('debugInfo', reference, timeleft,
     breaknumber, postponesnumber, settingsFile, logsFile, doNotDisturb)
@@ -1060,7 +1247,12 @@ ipcMain.on('open-preferences', function (event) {
 })
 
 ipcMain.on('set-contributor', function (event) {
+  const dir = app.getPath('userData')
+  const contributorStampFile = `${dir}/stamp`
+  const { DateTime } = require('luxon')
+  require('fs').writeFile(contributorStampFile, DateTime.now().toString(), () => {})
   global.shared.isContributor = true
+  log.info('Stretchly: Logged in. Thanks for your contributions!')
   if (preferencesWin) {
     preferencesWin.send('enableContributorPreferences')
   }
@@ -1087,7 +1279,7 @@ ipcMain.on('open-contributor-auth', function (event, provider) {
     backgroundColor: 'whitesmoke',
     webPreferences: {
       preload: path.resolve(__dirname, './electron-bridge.js'),
-      nodeIntegration: false
+      enableRemoteModule: true
     }
   })
   myStretchlyWindow.loadURL(myStretchlyUrl)
@@ -1103,14 +1295,11 @@ ipcMain.on('open-sync-preferences', function (event) {
 })
 
 ipcMain.handle('current-settings', (event) => {
-  return settings.data
+  return settings.store
 })
 
 ipcMain.handle('restore-remote-settings', (event, remoteSettings) => {
-  settings.restoreRemote(remoteSettings)
-  setTimeout(() => {
-    app.relaunch()
-    log.info('Stretchly: relaunching app')
-    app.exit(0)
-  }, 1000)
+  log.info('Stretchly: restoring remote settings')
+  settings.store = remoteSettings
+  initialize(false)
 })